test: move some files around
